Reduce duplication in config classes
The official COD services use a slightly different config than existing
prototype apps. Factor out the common base functionality into a
superclass and extend it.
